Traktor: Using "loose octree" to prevent splitting triangles which reduce geometry a lot.
[git-p4: depot-paths = "//depot/traktor/": change = 6040]

+++ code/Model/Utilities.cpp
@@ -759,7 +759,7 @@ void cullDistantFaces(Model& model)
 	}
 	model.setPolygons(polygons);
 
-	log::info << L"Culled " << (originalCount - polygons.size()) << L" polygon(s)" << Endl;
+	log::info << L"Culled " << uint32_t(originalCount - polygons.size()) << L" polygon(s)" << Endl;
 }
 
 	}
